pyproject.toml
[project]
name = "2026"
version = "0.1.0"
description = "Personal 2026 tracking system with Garmin sleep/wake automation"
requires-python = ">=3.11"
dependencies = [
    "garth>=0.5.0",
    "PyGithub>=2.1.1",
]

[project.scripts]
garmin-sync = "scripts.fetch_and_post:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
]
